requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.112.0
fastjsonschema==2.20.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
idna==3.7
importlib_resources==6.4.3
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.0
limits==3.13.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
mysql-connector-python==9.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
selenium==4.23.1
sentence-transformers==3.0.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
